+++ prepareTransaction/shellScript/holacontrato.sh
